+++ src/Transactions/CreateTransaction.cpp
@@ -1,4 +1,4 @@
 #include "CreateTransaction.hpp"
 
-CreateTransaction::CreateTransaction(const std::string &userName, UserType userType, float credits)
+CreateTransaction::CreateTransaction(const std::string &userName, UserType userType, double credits)
     : BasicTransaction(kTransactionType_Create, userName, userType, credits) {}
