+++ src/Transactions/LoginTransaction.cpp
@@ -1,4 +1,4 @@
 #include "LoginTransaction.hpp"
 
-LoginTransaction::LoginTransaction(const std::string &userName, UserType userType, float credits)
+LoginTransaction::LoginTransaction(const std::string &userName, UserType userType, double credits)
     : BasicTransaction(kTransactionType_Login, userName, userType, credits) {}
